backend/pyproject.toml
[project]
name = "plngx-dissect-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiofile>=3.9.0",
    "aiofiles>=24.1.0",
    "aiohttp==3.12.15",
    "apscheduler==3.10.4",
    "asyncstdlib==3.12.0",
    "diskcache>=5.6.3",
    "diskcache-stubs>=5.6.3.6.20240818",
    "fastapi==0.110.1",
    "flufl-lock==8.2.0",
    "jinja2==3.1.2",
    "pdfplumber==0.11.5",
    "pydantic==2.11.9",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.0",
    "pyyaml==6.0.3",
    "uvicorn[standard]==0.24.0",
]

[dependency-groups]
dev = [
    "austin-dist==3.7.0",
    "austin-tui==1.3.0",
    "austin-web==0.3.0",
]
